Revert changes to PsyTdpFnWPb
I had mistakenly changed PsyTdpFnWPb to not use ‘CalledFrom’ however it
is still needed.

At the same time I changed the routine name string to be a static
std::string const in the routine. This only creates the string a single
time and not every time the method is called.