Change output dir for Fortran targets
This makes the output dir for Fortran targets match the
other targets, which is build/Products.  Note that for
MinGW builds there will be no Debug/Release subdir for the
Fortran targets even when building through VS.

Also fix an error where ExpandObjects was not properly
located when using the VS / MinGw comobo.

[#69346724]