Avoid writing to sqlite db before it is initialized
Some fatal errors attempt to write to sqlite database before it has been
initialized.  This caused an access violation, so a check was added that
the pointer to sqlite database is initialized before accessing.

The fatal error that exposed this problem was in fact an error
reporting that the database failed to open!