[#62470514] Tell AWS code to not mark files public
With this commit the regressions tests will be uploaded to the S3
energyplus bucket, however, the files will not be public, so it's not
possible to view them.

Two possible outcomes exist:

1. The energyplus bucket is set to default public and these files will
become available
2. The user will be updated to be have permissions to mark files as
public.

If the case that #2 happens, we will need to update this flag to 'true'

@kbenne @Myoldmopar - note that with this commit regressions should be
fully working, and pushing to S3 you just aren't allowed to see the results.

@nllong do you have any advice for how to move this issue forward?