Add function to download and install files from web
Some parts of the EnergyPlus package are only available as pre generated
files from an external and manual process.  In these cases we have placed the prepared items in
EnergyPlusBuildSupport for assembly into the final EnergyPlusTeam
package.  This commit provides a new function install_remote that
downloads a file from a specified url and installs it much like the
internal cmake install command does. One caveat is that
EnergyPlusBuildSupport requires authentication so either the files need
to reside somewhere else, build support needs to be public, or a way to
authenticate (api key?) needs to be introduced.  This commit contains an
exmple of using the install_remote by pulling a file from the public
OpenStudio repository.  That example obviously needs to be replaced.

[#74735730]