Pass cmake version variables to transition subproject
On windows (without ifort) the transtion programs are a separate external
project nested inside the VS project.  Inside the subproject the cmake
version info variables are used.  This change makes sure those variables
are passed down to the inner project.  This was creating a problem where
the transition program was looking for the wrong idd filename on windows
builds.

[#74735730]