Connections to heat balance for cooling panel
These are the final draft changes to the code for the simple cooling
panel.  Most of these changes make the connections in the heat balances
and the zone equipment section.  Also fixed an improper name and fixed
some missing variables for other models in the Window complex and
equivalent layer code.  These might result in changes to output in
other input files.