Make EnergyPlusLib shared by default
This work makes EnergyPlusLib a shared library, with main becoming a
very light wrapper that calls into the function EnergyPlusPgm.
EnergyPlusPgm is the main function in the shared library's interface.

The EnergyPlusLib shared library is installed alongside the main
energyplus executable and distributed in the normal package.  In this
way there is no separate distribution of the EnergyPlus DLL.  It will
come with the regular installer package.