Remove DDOnly check in SQLite
I think this check is only valid for DDOnly environmental variable.
When I run an IDF, such as the RefBldgs, which run for only sizing
periods and not weather file run periods, they output an SQLite
database.

The main reason for this switch is so the CI machines will output
SQLite databases and there will be code coverage in the integration
tests without the need to force annual simulations.