Fix compile error and fix non-sqlite unit tests
There was a compile error with the merge of develop into this feature
branch.

Now all sqlite calls are wrapped by ‘if ( sqlite ) {}’ so non-sqlite
unit tests don’t need to instantiate sqlite to run unit tests that may
use outputs. This was causing errors in those other unit tests.