Configure IDD next to executable accounting for config dir
From an IDE build there is typically a Release/Debug sub directory under
Products.  The IDD was generated into Products, not accounting for the
config subdirectory.  This was a problem because the energyplus
executable expects the IDD to be located in the same directory.

A solution was implemented using cmake "generator expressions" to place
the IDD in the proper config subdirectory from IDE builds.  Furthermore
idd install commands were updated and the integration test runner was
updated to account for this new location.

CMake generate expressions were not supported in install commands until
cmake version 3, so the minimum required version of cmake has been
incremented to 3.  This may cause developers and/or robots to require
updating cmake!  Please consider this fact when reviewing.