Bug fixes
Due to a scoping issue with errorStream, there was a segfault in
certain situations. It is odd that the integration and regression
testing on the CI machine didn’t catch this. The errorStream is now a
shared_ptr owned by the SQLite instance (due to std::move())

Due to another scoping issue with strings in Schedules and Surfaces,
there was invalid data entered into the database. I thought these
strings were stored in a data structure in EnergyPlus but they are both
instead returned strings from a function causing the scoping issue. I’m
using std::move() here to avoid copy construction of new strings.