revise logic for when hard value for loop max flow
Only use user-specified value if it is smaller than the result from
non-coincident component flow requests.  User may throw in a large max
hard value and not intend it to make pumps and such huge.  But they may
use a small value and it needs to cap the design size