Convert from pointer to reference
This solves the malloc error i’m seeing on Mac. The issue is that the
pointer was never allocated memory on the heap and when EnergyPlus
exits, it tried to free unallocated memory.

This now uses references to achieve the same goal as the pointer,
however, there now needs to be an explicit constructor for SizingLog
that sets p_rVariable on construction.