[#84696534] Update transition fortran from develop
This branch was created from the v8.2.0 tag so that we can show the
reports are the same as the official v8.2.0 validation reports.

However, I was unable to get the transition fortran projects to build
without pulling in the transition code from current `develop` so this
particular commit should probably not be merged with develop when it's
time for that.