Make embedded images locations properly relative
I expect at least @axelstudios to have an opinion on this, but it should
make the documents at least partially browseable from inside of github.

Also, it just makes more sense. The paths are incorrect with the current
file layout.

I am willing to revert this commit if there's a reason to.

[#84691154]