Add InputProcessor unit tests
This will currently read in an IDD snippet or the Full Energy+.idd file
in the Products folder and process it. The fixture takes care of all
deallocations in the TearDown() phase so no extra steps are needed in
the unit tests themselves.

This still needs some work. The InputProcessorFixture should probably
be in its own header file OR this should go in the EnergyPlusFixture
since this usage will probably be common between a lot of different
unit tests.