Implement IDD cache, show message, clean up
This implements IDD caching since it is now part of EnergyPlusFixture
SetUpTestCase(). These means it will be run once when the Fixture is
setup before any test cases are run based on that fixture. I did this
because I eliminated the option to pass an IDD snippet to processIDF
and made processIDD private. I was worried that the IDD snippets would
be incorrect or become out of date as Energy+.idd is updated. So the
full IDD is now read unless the cached IDD is used.

One thing is that the IDD is parsed for caching for any Fixture that
inherits from EnergyPlusFixture as well. This happens by default IF
static void SetUpTestCase() {} is not part of the derived class. If
that code is in the derived class then it will not initialize the
m_idd_cache. This is important to know because if you need to parse an
IDF then you need to add EnergyPlusFixture::SetUpTestCase(); to that
function like so: static void SetUpTestCase() {
EnergyPlusFixture::SetUpTestCase(); }

I also typedef’d the fixtures as FooDeathTestFixture so that we can
have separate death test only unit tests but they retain everything
about the fixture.

I also created a show_message() function that grabs the
test_case_name() and name() from the unit test and creates the
ShowMessage text that was in each unit test. This gets called every
time EnergyPlusFixture or any derived fixture is used. This eliminates
the need to put this text in every unit test as well as automatically
updates every time the unit test name changes.