This is a major clean up and adds a custom parser
The parser is needed for the IDF snippets since it now looks for all
required objects in a snippet. So instead of making everyone add
Building and GlobalGeometryRules, it now looks for those objects and
adds them if they are not in the IDF. Otherwise it uses the existing
objects. The parser is a recursive descent parser so it is quick.

I started to add documentation to the new functions. There needs to be
more still.

I also still need to fix a few lingering issues that will help
developers to use this more easily.

There are also now unit tests for the EnergyPlusFixture under
EnergyPlusMetaFixture (so meta…)