Add clear_state(), move function static, clean up
First, this adds a clear_state() function to various namespaces. This
function will do as it says, clear the state of the namespace. This is
important for unit testing. This function should never be used during
normal simulation. This functionality is basically moved from
EnergyPlusFixture, but having it in the namespace allows for the next
change.

Moved various InputProcessor function static variables into the
namespace. However, for both InputProcess and OutputProcessor these
variables are not in the header file as ‘extern foo’. This means no one
outside of that implementation file can use those variables. This
maintains some of the original intend. However, this means those
variables can now be reset via the clear_state() function.