Fix Package hyperref Warning: Suppressing empty link ...
This was caused by a pattern of \protect\hyperlink{section-1}{} getting inserted where there was a [] in the original text.

Fixed by a global search and replace to {[]}

Note: this commit also includes a few minor edits such as adding hyphenation to long EnergyPlus class names as well as replacing greek letters with their equivalent in mathmode.