Restore missing sections from Engineering Ref Variable Refrigerant Flow Heat Pumps
These missing sections were noticed during a verification check comparing the headings
of the original markdown with the headings of the generated LaTeX. There appears to have
been an issue with a regular expression that was used to transform the document -- that regex
took more content than it should have and only processed part of it, resulting in missing
content for the generated LaTeX code.

These sections have been restored by hand from the original Markdown.