ExpandObject - Missing Setpoint
When the HVACTemplate:System:DedicatedOutdoorAir object is used with no
cooling coil, a heating coil and is setup with heat recovery, a severe
error occurs. A setpoint manager is missing at the reference setpoint
node for the economizer setpoint manager (SetpointManager:MixedAir).

The change to epfilter.f90 seems to fix the issue. I've recompiled
ExpandObjects locally and tried it with EnergyPlus v8.5.0 and my model
runs fine now.

It hope this description is clear enough!