Re-correct file indentation
It was noticed that when the Advanced Surface Concepts was restored, the two sections under it ("Exterior Naturally Vented Cavity" and "Green Roof Model ...") were at the same level as Advanced Surface Concepts. The actual problem was that Advanced Surface Concepts was restored as a section as opposed to as a chapter (which it is a chapter according to EnergyPlus docs 8.4). Thus the indentation of exterior-naturally-vented-cavity.tex and green-roof-model-ecoroof.tex has been restored and the Advanced Surface Concepts section delimiter has been changed to "\chapter".