Addition of unit tests
Two unit tests were added to verify that the code work in the movable
insulation is pulling the correct material properties.  For non-glazing
materials, it uses simply the solar absorptivity.  For glazing
materials, it subtracts the transmissivity and the front side solar
reflectance from unity to calculate the solar absorptivity.