Bug fix and doc update
This commit includes the code to fix this bug and a slight modification
to the documentation in the IO Ref to clarify what EnergyPlus is doing.
 As per discussion with Mike Witte, the error call has been changed to
severe and the user is required to account for all radiant energy
(within a small tolerance).  The error message produced when this does
not happen has also been improved to provide the user with more
information about the problem.