Fix of defect
The user specified location temperature was not being calculated for
low temperature radiant systems.  This fix implemented the calculation
of this value.  Testing included asking for the temperature at the
source location and comparing it to the temperature already calculated
at that location. This got the exact same result for all time steps.
When the location of the user specified temperature was changed, a
different value between the source location and the zone air
temperature was obtained.  This value was noted to vary in how close it
was to the other values based on whether or not the radiant system was
on.  Both of these tests were successful.