Changes to address speed issue
The implementation of this defect fix ended up resulting in a 0.4%
speed increase in the test suite.  Many changes were made to try to
eliminate this speed issue while still maintaining all of the
functionality.  Now, variables won’t be calculated, initialized, or
even allocated if they won’t be used.