Add cache map to reduce complexity
This reduces the algorithmic complexity of GetFooInput (really
GetObjectItem) from O( n ^ 2 ) to O( n log n ). This has a dramatic
effect on runtime for GetObjectItem function. We initially thought
incrementing an iterator would be constant time operation but since
this is a map, it is actually linear. This commit changes it to log
time lookup.