Fix of crash
This fix avoids the crash that discovered when a high temperature
radiant heater is autosized using scalable sizing factors as in the
defect example idf.  This needs to be looked into for other equipment
that uses the scalable sizing process to make sure that this is not an
issue elsewhere.  Committing this simple fix to trigger the CI runs to
make sure it doesn’t introduce a new problem.