Unit test verifying correct assignment of flag
This new unit test verifies that the DataScalableSizingON flag is rest
properly.  It was being left to true and this was causing a crash under
certain equipment combinations.  The original input file that caused
the problem combined high temperature radiant heaters with DX coils.
Now, all zone type equipment has this flag set back to false once the
sizing process is complete.