Move UA Calculation
Moved the UA calculation to be in a similar place as the baseboard
models.  The placement in the Init routine caused all sorts of problems
(variable values of NaN) because some variables had not been calculated
yet.  Moving this to the sizing routine corrects the problem and makes
a bit more sense.