Removed unnecessary VerifyName()
In some parts of the source file we had unnecessary VerifyName() calls,
since json() takes care of duplicates we could remove it. There is 1
place in ZoneTempPredictorCorrector.cc where we have to verify for
duplicates among Names and ZoneNames, and another place where we have
to verify for duplicates among SingleCooling and SingleHeatingOrCooling
objects. I kept Verifyname there. Search algorithm might needed to be
re-thought.