Fix failing unit tests
Still failing:

EnergyPlusFixture.PlantHXModulatedDualDeadDefectFileLo
EnergyPlusFixture.CondenserLoopTowers_SingleSpeedSizing

The first one has an active issue still. The second one is surprising.
I thought it was due to hardcoded node number and I switched to using a
lookup for the proper index but it didn’t fix it. There is probably
some other underlying issue.