Restore deleted section header for "Group Advanced Surface Concepts"
In commit 25ce9f90474ca6145dd2b43836fd90262e450e3b on 2016-07-06, the
author was apparently trying to get the file
"group-advanced-surface-concepts.tex" to compile stand-alone. They had
deleted the section header and added a latex prelude which was
accidentally committed. In undoing that commit, the section header was
mistakenly deleted. This fix restores the section header to the file.