Modification to moisture balance
Code in predictor/corrector was modified to correct account for the
humidity ratio of the air entering the zone from the earth tube as it
might be different than the outside air.  Also added another variable
to the input files in the test suite so that the impact of this change
can be tracked.