Eliminate Negative Baseboard Heating
This commit eliminates the problem noted where negative heating
(LoadMet) was being reported for the baseboard system.  In addition to
checking the logic of the “ZeroSource” variable, this also cleaned up
some of the assignments surrounding when the system was on and off.
Unit tests are also included on this commit.