Switch documentation from nonstopmode to batchmode
Change the documentation building flag in CMake from

"-interaction=nonstopmode" to "-interaction=batchmode"

Both do not halt on warnings or errors but the second flag will not
print out to the command-line. Since LaTeX gives us the build logs for
error reporting, the command line print-out is redundant but I realize
it may be useful in conjunction with the CI.

I did a highly informal test of two cases with the current flag and the
batchmode flag on my mac and build times are slightly faster with the
"batchmode" flag:

-interaction=nonstopmode

real	2m56.202s
user	8m37.940s
sys	0m14.456s

-interaction=batchmode

real	2m32.310s
user	7m42.362s
sys	0m10.172s

Potential downsides are that CI may want to process error messages from
stdin. Could a python script be added that checks the LaTeX logs for
errors and reports a succinct list of issues such as errors and
"multiply defined labels"?