Clear std::cin and std::cout from being put into an error state
Standard input and standard output are placed into the ObjexxFCL global streams. Apparently, during iteration over these files via the FindUnitNumber(...) function in UtilityRoutines.cc, calls to gio::inquire(Unit cosnt unit, IOFlags &flags) which calls stream->size() cause these streams to be placed into an errored state.
It may be better that standard in and standard out are removed from the global streams listing. However, ObjexxFCL has been notified of the issue and they are considering changes for the next ObjexxFCL release that will fix the problem.

In the meantime, an easy way to handle the situation is to check if cin (standard in) or cout (standard out) are in an errored stae (i.e., not good) and clear their error flags if so.

The routine has been updated to clear cin/cout error flags if needed and to also print out to both cerr and cout to demonstrate cout is working.