Unit Ventilator Documentation Changes
Documentation changes to accompany the code changes.  Variable percent
for both with and without a cooling coil while cooling agree with the
code as modified for this bug fix.  Note that above these sections in
both documents the text was wrong before but is correct now without
changes to the text because it is in line with the modified code.