Solution allows GlassEquivalentLayer to be used
This change allows GlassEquivalentLayer materials to be used to define
Movable Insulation.  The calculation code for movable insulation
previously allowed this, but the get routine blocked this material from
being used for this.  This corrects that and adds a new required
thermal resistance and a default value that is needed by the simulation.