Unit test, new input file for test suite, correction
This commit includes a unit test for a new function that was added as
part of the bug fix.  In addition, there is a new test suite file that
was added to exercise the cycling fan option for the unit ventilator.
Finally, a correction was made to the function as it was using a Real64
variable instead of what it should have been (int).