Addition of check for no mass or zero thickness materials
Site:GroundDomain:* objects require insulation but using a no mass
material for the insulation causes EnergyPlus to crash.  This code adds
a trap on reading input to flag this problem and force the user to fix
this.  This check needs to be made in four different places so the fix
includes two generic simple subroutines that check to see if there is a
problem and then throw an error message if there was a problem.
Testing indicates the error message is being produce when it should and
not being sent to the err file if things are fine.