Addition of reference/class to IDD for Cooling Panel
This adds two lines to the IDD that will allow cooling panels to be identified as equipment on branches.  This already works in EnergyPlus as attested to by four existing input files in the test suite.  However, by correctly defining this, it would allow interfaces to properly assign this as well.