Exit without simulation if using convert flag
BREAKING CHANGE

This changes the existing behavior of the convert flag, which currently continues the simulation. Users wanted a way to only convert files from IDF->epJSON and epJSON->IDF without running a full simulation. This allows that functionality while breaking the existing workflow. I think this is an appropriate tradeoff.