Fixed the custom meter issue
NREL#7655
The issue was also reported in the helpdesk ticket# 15343.
The function GetCustomMeterInput (), which processes the output variables for custom meters, ran before the function AllocateAndSetUpVentReports(), which sets up many output variables.
The issue was fixed by moving the function UpdateMeterReporting(), together with the function verifyCustomMetersElecPowerMgr(), after the function CreateEnergyReportStructure().