Correction of hard crash in pool model
This commit fixes the hard crash that takes place when the user chooses a pool surface that does not exist in the input file.  This results in a surface pointer of zero which then causes a bad indexing problem because the index is zero.  This is now fixed and a new separate subroutine was created for the various checks made on the pool surface.