Refactoring of Main Simulation Routines
This commit looks at refactoring the main simulation routines for the various low temperature radiant system types.  There is potentially other things that could be done at this point, but this is a good "intermediate" point.  More work could be done restructing and splitting apart existing routines, though whether or not that happens has not yet been decided.