Rebug the code
This commit unfixes a "bug" that was discovered during the implementation of on-off control for radiant systems.  The bug is that the variable flow radiant system the old code allowed for the maximum flow fraction to be greater than 1.0.  Unfixing this to see if the surprising number of output file diffs were caused by this fix.