Unit test and modified IDF/RVI files
Added a unit test to look specifically at the limit when the flow/power fraction is greater than 1 (not allowed--limited to 1.0).  Also modified IDF and RVI files to produce output for the system mass flow rate.  This was done because this value specifically has been corrected with the fix that was made.