Fix getAbsolutePath to resolve mac issue.
On mac, `getProgramPath` :
```
_NSGetExecutablePath, /Users/dev/git/EnergyPlus-build-release/Products/energyplus, pathSize=1024
```

This is a symlink, but points to a relative dir
```
ls -la Products/energyplus
lrwxr-xr-x 1 dev staff 16 Mar  1 17:55 Products/energyplus -> energyplus-9.5.0
```

So witht he previous implementation, fs::absolute is called after, and that resolves to current_dir / energyplus-9.5.0. So unless you called the call from within the exe directory, it failed. THis fixes it