Reduce dynamic allocations discovered with heaptrack
File tested: BenchmarkHospitalNew_USA_CA_SAN_FRANCISCO.idf

 * Dynamic allocations reduced from 38,231,848 to 5,569,555

Moving from `static const std::string RoutineName` to
`constexpr auto RoutineName` is a good idea, however it ended up
resulting in extra conversions from `const char *` to `std::string`
which is silent.

`static const std::string` has in inherent cost in that each
time the function is entered a thread-safe check must be performed.

Similarly, constructing a new `std::string` from a `const char *`
is going to have potential dynamic allocation cost, which is why
we did not notice a performance degredation when the change was made.

This patch set

 * moves from `constexpr auto RoutineName` to
   `constexpr static std::string_view RoutineName`
 * avoids automatic conversions to `std::string` because that
   does not exist for `std::string_view`
 * similarly to how `const` makes you think about "`const`-correctness",
   `std::string_view` pushes a further set of changes

Because of the move to `std::string_view` we needed to update the
function signatures of a few functions that do logging with
RoutineName.

Because of the move to `std::string_view`, which does not have a `+`
operator, we then needed to modernize some of the logging function
calls to use `fmt::format`.

For any future update of `const static std::string`, I recommend
a move to `constexpr static std::string_view` and updating
all downstream code to avoid as many dynamic allocations
as possible.

The benefit is clear, with a measurable performance improvement
and potentially drastic decrease in dynamic allocations.