Introduction of solar cosines
Three different places in the code there was cases of BeamSolarRad being used without modification.  This variable is the amount of solar perpendicular to the current direction of the sun, and the sun is not always perpendicular to the ground or a roof.  These cases were corrected in the code to use the appropriate solar cosine (SOLCOS(3) was used for ground models whereas the ecoroof model used solar angles based on the tilt and orientation of the actual roof).