Correct typo in LowTempRad:VarFlow input
There was a bad cut and paste/typo in the input for the variable flow radiant systems.  When a cooling only system was input by the user and autosizing used, a severe error would be thrown because E+ was looking at the wrong variables.  This has been corrected and a unit test added to check that this no longer throws a severe error.  No documentation changes needed or changes to the test suite idfs.