Remove local copy of Eigen library
Based on @jmarrec's recommendation, the Eigen library located in
EnergyPlus's third party folder will be referenced rather than
including a shrunk copy of Eigen with this library. Even with the
reduced library that was being included, there were many files. If
this manages to build on all OS's successfully, it will make the merge
diffs look much nicer and more tractable.