Water Tank Underflow Fix
Correction of the problem.  There were a variety of underflow issues in the water tank model as well as a potential divide by zero.  These were not causing a hard crash but eventually led to a fatal error that stopped EnergyPlus.  The solution stops the problem at the source to void sending incredibly tiny flow values to other routines and leading to problems.