Water Thermal Tank Negative Electric Fix
Water Thermal Tank was reporting negative electricity.  This was eventually traced back to the backup heating element being set before autosizing took place.  As a result, it was set to -99999 (DataSizing:AutoSize value) and this led to negative electric consumption.  This commit correct the problem by not allowing the backup element capacity to be set if autosizing has not yet been done.