8154 Changes to View Factor after Review Comments
As a result of comments made by @jcyuan2020, the method for figuring out view factors to ground and sky when the calculated view factor to shelf causes the sum of all three to be greater than zero was completely changed.  The method is now more robust and follows the suggestions with some additional features as well.