Elimination of unused parameters
This commit eliminates the two parameters that are not used in the FuelFactors input.  This includes code changes, IDD changes, IDF changes, transition updates, and one documentation change.  Still yet to do: create a unit test and add a table to the IO Ref for source to site ratios from EPA.