Revised ASHRAE Tarp Correction
Modified, much simpler approach to eliminating the cosTilt = 0 case from the ASHRAE Tarp Natural Convection model.  The cases where cosTilt is exactly zero does not warrant the additional case because the reduced and enhanced models yield the same result and because the times when cosTilt is literally zero without any trailing junk is probably limited.  A new unit test and some engineering reference changes are included.  Left the "vertical" correlation in the code because it is used by other things.