Correct CondFD Error
User had an input file where an air layer was defined with a regular material and then a construction that did not use that layer used the CondFD solution.  CondFD looked at everything and then produced an error message because the air layer didn't fit the CondFD requirements.  This has now been fixed and CondFD will only look at constructions that actually use the CondFD solution.  A new subroutine and unit test are included.