Window5 Data File SC/SHGC 0 Fix and Unit Test
When Window5 outputs a file that has a SC and/or SHGC of 0, EnergyPlus has been throwing a severe error.  Problem is that it never uses those values for anything because the model inside EnergyPlus is not based on those parameters but more detailed information found elsewhere in the Window5 data file.  So, it does not make any sense to severe error out for this.  The error is now just a warning message and the program continues without a problem.  A unit test to verify the warning messages when this happens is also included in this commit.