Correction of OASetPtMgr Schedule Value Issue
User noted that the schedule value for the outdoor air reset manager is only supposed to be 1 or 2.  The code, however, accepted any value and then defaulted to 1.  This has been corrected so only 1 or 2 is allowed.  Anything else produces an error message.  At the get, if looks to see if the schedule is less than 1 or greater than 2.  At the calculation, if looks to see if it's something other than 1 or 2.