Correction to Reporting Problem
The code was skipping partitions for updating the outside face flux in the heat balance because there is no outside face--the surface is adiatic.  However, the code to skip that forgot that interzone partitions do need to figure this out.  This was noticed by a user and the code before the fix resulted in output that made it look like the surface was accumulating energy gain/loss for interzone surfaces over an annual simulation eventhough this value should be relatively small.  With the fix, the outside face flux is now no longer zero all the time and the summation of heat storage in the surface is more appropriate as well.