Back out changes made to code
After discussion during review of what was done, it was decided to go in a different direction.  This commit undoes almost all of the changes other than a few minor, random documentation fixes.  This will provide a new base from which to start for the new solution to this defect.