Fix for Flow/Energy Inconsistencies
Swimming Pools under certain situations (no bypass) were producing results that were invalid.  Conservation of mass was not being preserved across the pool and neither was conservation of energy.  This was due to the swimming pool setting flows at the outlet but continuing to accept the inlet flow.  As a result of this, the inlet mass flow did not equal the exit mass flow.  In fact, in the defect file, it wasn't even close.  After much testing with different solutions, this solution fixed the situation where no bypass was present and did not seem to impact situations where a bypass was present (those results were fine already).