Revised solution
Revision to the solution previously proposed that wasn't actually working.  This works now and only adds to the simulation time when a pool is present.  The mass flow rate across the pool is now constant, even if a bypass is present.  So, the pool is now more like an on-off with full flow.  This was tested with the user defect .idf, a single branch demand loop variation of the user file, and a variation where a bypass was put in parallel with the pool.  All operate as constant flow/on-off control.