Solution to CondFD Flux Imbalance
It was determined that the outside surface flux was wrong for a multi-year simulation.  Testing and comparing two output variables that should be almost identical showed than when the IsRain flag was true that things were way off.  This is because the one variable was not being calculated properly.  As a result, there were flux inconsistencies.  The fix corrects this problem and eliminates the inconsistency.  A unit test is also included in this commit.