Fix to problem of zero capacity
In the variable speed coils, if the user entered a zero capacity a crash resulted in the simulation.  A previous solution looked for a valid capacity among the stages and used that.  This was flagged as undesirable during review because reviewers wanted a zero capacity to simply eliminate the coil.  This solution does that and fixes things so that the coil just doesn't ever run and avoids the crash that was happening before the fix was applied.