Fix of problem
When there are no internal gains are present and VRF is used, E+ crashes because some structures are not allocated before it gets to the VRF code.  Code was added to resolve this issue by getting the zone equipment data.  Other examples of this strategy already exist in the code.