Moving GetZoneEquipment on up
The call to GetZoneEquipment has been moved up to the HeatBalanceManager.cc and called from GetBuildingData right after the call to GetZoneData.  All other calls in the code to GetZoneEquipment have been removed.  Various unit tests were altered to adjust for this new reality.  Now, only one call is ever made and other models do not have to patch in a call.