Addition of unit test
Added a unit test.  It runs GetZoneEquipment and properly resets the get flag.  Then the second call to GetZoneEquipment, it doesn't do anything which is correct.  If it goes into GetZoneEquipmentData which is called by GetZoneEquipment, the program crashes which would be correct.  So, this tests that only one time into GetZoneEquipmentData is allowed.