Solution to problem
This fix institutes the use of the glycol routines which are used elsewhere in the plant simulation.  In addition, this is now more efficient since this constant value is only calculated once avoiding numerous calls to the glycol density routine.  Volumetric flow rates in component and in node data now agree in the .idf file that spurred this defect.