Fix for GenContam issues
There were a few places where GenContam was not implemented properly in the code.  After a thorough search of the code for both CO2 and GenContam, it was discovered that there were a couple of places where the code for these two entities were not similar (GenContam code that needed to be present was missing).  Also, an incorrect assignment of the simulation flag was also discovered.  This commit fixes those issues.